requirements.txt
asgiref==3.8.1
Django==5.1.1
django-filter==24.3
django-restframework==0.0.1
mysqlclient==2.2.4
sqlparse==0.5.1
typing_extensions==4.12.2
